# Speed up multiple nibble searches