@BurntSushi 2019-07-07
This change is in `memchr 2.2.1` on crates.io.